# alzawaj-project-backend/src/config/openapi.yaml
openapi: 3.0.0
info:
  title: Zawag Islamic Marriage Platform API
  version: 1.0.0
  description: API documentation for the Zawag Islamic Marriage Platform
servers:
  - url: http://localhost:5001
    description: Development server
  - url: https://alzawaj-backend.onrender.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
          example: "6051d2d7e1b8a20015d23f5f"
        chatRoomId:
          type: string
          example: "6051d2d7e1b8a20015d23f5e"
        senderId:
          type: string
          example: "6051d2d7e1b8a20015d23f5a"
        content:
          type: string
          example: "Assalamu alaikum! How are you doing today?"
        status:
          type: string
          enum: [pending, approved, rejected, flagged]
          example: "approved"
        createdAt:
          type: string
          format: date-time
          example: "2021-03-17T10:30:00Z"
        approvedAt:
          type: string
          format: date-time
          example: "2021-03-17T10:31:00Z"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "بيانات التسجيل غير صحيحة"
        statusCode:
          type: integer
          example: 400
        error:
          type: array
          items:
            type: string
          example:
            - "تأكيد كلمة المرور مطلوب"
            - "الاسم يجب أن يكون بين 2-100 حرف"
            - "العمر يجب أن يكون بين 18-100 سنة"
            - "حالة اللحية مطلوبة للذكور"
            - "الحالة المالية مطلوبة للذكور"
            - "نوع السكن مطلوب للذكور"
security:
  - bearerAuth: []
...
paths:
...
  /api/admin/messages/pending:
    get:
      tags: [Admin]
      summary: Get pending messages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      messages:
                        type: array
                        items:
                          $ref: '#/components/schemas/ChatMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/messages/{messageId}/approve:
    post:
      tags: [Admin]
      summary: Approve message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Message successfully approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/admin/messages/{messageId}/reject:
    post:
      tags: [Admin]
      summary: Reject message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Message successfully rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/reports/{reportId}/action:
    post:
      tags: [Admin]
      summary: Report actions
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [assign, resolve, dismiss]
                notes:
                  type: string
      responses:
        '200':
          description: Report action successfully processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...